# movie-api-yaml/swagger.yaml
openapi: 3.0.0
info:
  version: 'initial'
  title: 'MovieDB API'
  description: 'A Movie API for EWD 2020'
paths: 
  /api/movies:
    get:
      operationId: list-movies
      summary: List Movies
      description: List movies from movie db.
      responses:
        '200':
          description: 'List of Movies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/movieListObject'
              examples:
                movieListObject:
                  $ref: '#/components/examples/movieListObject'
        '404':
            description: "401 Error"
            content: 
              application/json:
                schema:
                  type: object
                  example: {message: "unable to find Movies", code: 404}
        '500':
           description: "500 Error"
           content: 
              application/json:
                schema:
                  type: object
                  example: {message: "Internal Error", code: 500}
                  
    post:
      operationId: add-movie
      summary: Add a Movie
      requestBody:
        description: Movie data as JSON object  
        required: true
        content:
          application/json:
            schema:
              type: object
              example: { id: 181808, imdb_id: "tt2527336", title: "Star Wars: The Last Jedi"}
      responses:
        '201':
          description: 'Successfully added new Movie'
          content:
            application/json:
              examples:
                movieListObject:
                  $ref: '#/components/examples/movieObject'
        '405':
          description: "405 Error"
          content: 
            application/json:
              schema:
                type: object
                example: {message: "Invalid Movie Data", code: 405}
                
  /api/movies/{id}:
    get:
      operationId: get-a-movie
      summary: Get a movie by id
      description: Get movie by id
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Unique ID of the movie to get
      responses:
        '200':
          description: 'Movie details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/newMovieObject'
              examples:
                movieListObject:
                  $ref: '#/components/examples/movieObject'
        '404':
          description: "404 Error"
          content: 
            application/json:
              schema:
                type: object
                example: {message: "unable to find Movie", code: 404}
        '500':
          description: "500 Error"
          content: 
            application/json:
              schema:
                type: object
                example: {message: "Internal Error", code: 500}
                
    put:
      operationId: update-a-movie
      summary: Update a movie by id
      description: Update movie by id
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Unique ID of the movie to get
      requestBody:
        description: Movie data as JSON object  
        required: true
        content:
          application/json:
            schema:
              type: object
              example: { id: 181808, imdb_id: "tt2527336", title: "Star Wars: The Last Jedi"}
      responses:
        '200':
          description: 'Movie Updated'
          content:
            application/json:
              schema:
                type: object
                example: { id: 181808, imdb_id: "tt2527336", title: "Star Wars: The Last Jedi"}
        '404':
          description: "404 Error"
          content: 
            application/json:
              schema:
                type: object
                example: {message: "unable to find Movie", code: 404}
        '500':
          description: "500 Error"
          content: 
            application/json:
              schema:
                type: object
                example: {message: "Internal Error", code: 500}

    delete:
      operationId: delete-a-movie
      summary: Delete a movie by id
      description: Delete movie by id
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Unique ID of the movie to get
      responses:
        '200':
          description: 'Movie Deleted'
          content:
            application/json:
              schema:
                type: object
                example: { id: 181808, status: "deleted"}
        '404':
          description: "404 Error"
          content: 
            application/json:
              schema:
                type: object
                example: {message: "unable to find Movie", code: 404}
        '500':
          description: "500 Error"
          content: 
            application/json:
              schema:
                type: object
                example: {message: "Internal Error", code: 500}
# Added by API Auto Mocking Plugin
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AaronMooney/Movie/initial
    
    
components:
  schemas:
    movieListObject:
      title: Movie List Result Object
      type: object
      properties:
        movies:
          type: array
          items:
            type: object
            properties:
              poster_path:
                $ref: '#/components/schemas/imagePath'
              adult:
                type: boolean
              overview:
                type: string
              release_date:
                type: string
              genre_ids:
                type: array
                items:
                  type: integer
              id:
                type: integer
              original_title:
                type: string
              original_language:
                type: string
              title:
                type: string
              backdrop_path:
                $ref: '#/components/schemas/imagePath'
              popularity:
                type: number
              vote_count:
                type: integer
              video:
                type: boolean
              vote_average:
                type: number
    imagePath:
      title: image-path
      type: string
      default: null
      
    newMovieObject:
      title: New Movie Object
      type: object
      properties:
        poster_path:
          $ref: '#/components/schemas/imagePath'
        adult:
          type: boolean
        overview:
          type: string
        release_date:
          type: string
        genre_ids:
          type: array
          items:
            type: integer
        id:
          type: integer
        original_title:
          type: string
        original_language:
          type: string
        title:
          type: string
        backdrop_path:
          $ref: '#/components/schemas/imagePath'
        video:
          type: boolean
          
  examples:
    movieListObject:
      value:
        movies:
        - title: The Avengers
          poster_path: /cezWGskPY5x7GaglTTRN4Fugfb8.jpg
          overview: >-
            When an unexpected enemy emerges and threatens global safety and security, Nick Fury, director of the international peacekeeping agency known as S.H.I.E.L.D., finds himself in need of a team to pull the world back from the brink of disaster. Spanning the globe, a daring recruitment effort
            begins!
          release_date: '2012-04-25'
          genre_ids:
            - 878
            - 28
            - 12
          id: 24428
          original_title: The Avengers
          original_language: en
          backdrop_path: /hbn46fQaRmlpBuUrEiFqv0GDL6Y.jpg
          popularity: 7.353212
          vote_count: 8503
          video: false
          vote_average: 7.33
        - title: 'Avengers: Age of Ultron'
          poster_path: /t90Y3G8UGQp0f0DrP60wRu9gfrH.jpg
          adult: false
          overview: >-
            When Tony Stark tries to jumpstart a dormant peacekeeping
            program, things go awry and Earth’s Mightiest Heroes are put to the ultimate test as the fate of the planet hangs in the balance. As the villainous Ultron emerges, it is up to The Avengers to stop him from enacting his terrible plans, and soon uneasy alliances and unexpected action pave the way for an epic and unique global adventure.
          release_date: '2015-04-22'
          genre_ids:
            - 28
            - 12
            - 878
          id: 99861
          original_title: 'Avengers: Age of Ultron'
          original_language: en
          backdrop_path: /570qhjGZmGPrBGnfx70jcwIuBr4.jpg
          popularity: 7.557812
          vote_count: 3924
          video: false
          vote_average: 7.4

    movieObject:
      value:
        adult: false
        backdrop_path: "/5Iw7zQTHVRBOYpA0V6z0yypOPZh.jpg"
        genres: 
          - id: 14,
            name: "Fantasy"
          - id: 12
            name: "Adventure"
          - id: 878
            name: "Science Fiction"
          - id: 28,
            name: "Action"
        homepage: "https://www.starwars.com/films/star-wars-episode-viii-the-last-jedi"
        id: 181808
        original_language: "en"
        original_title: "Star Wars: The Last Jedi"
        overview: "Rey develops her newly discovered abilities with the guidance of Luke Skywalker, who is unsettled by the strength of her powers. Meanwhile, the Resistance prepares to do battle with the First Order."
        poster_path: "/kOVEVeg59E0wsnXmF9nrh6OmWII.jpg"
        release_date: "2017-12-13"
        title: "Star Wars: The Last Jedi"
        video: false
        vote_average: 0
        vote_count: 0